Create functions.py
Implement the starts_with() function taking two strings and returning true if the first argument starts with the second one.
Implement the ends_with() function taking two strings and returning true if the first argument ends with the second one.
Implement the swap_strings() function taking two strings and swapping their values.
Implement the find_last_not_of() function taking a string and a character sequence as its arguments. The function returns the last character equal to none of the characters in the given character sequence.
Implement the convert_to_int() function taking a string argument and returning its integer representation. Throw an exception if the conversion is not possible.
Implement the rotate_by() function taking an integer array and a number as its arguments. The function returns the array rotated by the specified number of positions (given by the second argument).